requirements.txt
tensorflow==2.12.0
flask==2.3.2
pillow==10.0.0
numpy==1.24.3
protobuf<4.0.0